+++ Howatworks.PlayerJournal.Serialization/Exploration/MultiSellExplorationData.cs
@@ -13,8 +13,8 @@ namespace Howatworks.PlayerJournal.Serialization.Exploration
         }
 
         public List<DiscoveredItem> Discovered { get; set; }
-        public int BaseValue { get; set; }
-        public int Bonus { get; set; }
-        public int TotalEarnings { get; set; }
+        public long BaseValue { get; set; }
+        public long Bonus { get; set; }
+        public long TotalEarnings { get; set; }
     }
 }
