+++ Howatworks.PlayerJournal.Serialization/Other/DatalinkVoucher.cs
@@ -2,7 +2,7 @@
 {
     public class DatalinkVoucher : JournalEntryBase
     {
-        public int Reward { get; set; }
+        public long Reward { get; set; }
         // TODO: check these
         public string VictimFaction { get; set; }
         public string PayeeFaction { get; set; }
