+++ Howatworks.PlayerJournal.Serialization/Other/Resurrect.cs
@@ -4,7 +4,7 @@
     {
         // TODO: confirm type
         public string Option { get; set; }
-        public int Cost { get; set; }
+        public long Cost { get; set; }
         public bool Bankrupt { get; set; }
     }
 }
