+++ Howatworks.PlayerJournal.Serialization/StationServices/MissionAbandoned.cs
@@ -8,6 +8,6 @@ namespace Howatworks.PlayerJournal.Serialization.StationServices
         public string Name { get; set; }
         [SuppressMessage("ReSharper", "InconsistentNaming")]
         public long MissionID { get; set; }
-        public int? Fine { get; set; }
+        public long? Fine { get; set; }
     }
 }
