+++ Howatworks.PlayerJournal.Serialization/StationServices/ModuleBuy.cs
@@ -9,7 +9,7 @@ namespace Howatworks.PlayerJournal.Serialization.StationServices
         public string Slot { get; set; }
         public string BuyItem { get; set; }
         public string BuyItem_Localised { get; set; }
-        public int BuyPrice { get; set; }
+        public long BuyPrice { get; set; }
         public string Ship { get; set; } // Note: ship type
         [SuppressMessage("ReSharper", "InconsistentNaming")]
         public int ShipID { get; set; }
@@ -18,6 +18,6 @@ namespace Howatworks.PlayerJournal.Serialization.StationServices
 
         public string SellItem { get; set; }
         public string SellItem_Localised { get; set; }
-        public int? SellPrice { get; set; }
+        public long? SellPrice { get; set; }
     }
 }
