+++ Howatworks.PlayerJournal.Serialization/StationServices/SellDrones.cs
@@ -4,7 +4,7 @@
     {
         public string Type { get; set; } // TODO: localised?
         public int Count { get; set; }
-        public int SellPrice { get; set; }
-        public int TotalSale { get; set; }
+        public long SellPrice { get; set; }
+        public long TotalSale { get; set; }
     }
 }
