+++ Howatworks.PlayerJournal.Serialization/StationServices/ShipyardBuy.cs
@@ -7,13 +7,13 @@ namespace Howatworks.PlayerJournal.Serialization.StationServices
         [SuppressMessage("ReSharper", "InconsistentNaming")]
         public long MarketID { get; set; }
         public string ShipType { get; set; }
-        public int ShipPrice { get; set; }
+        public long ShipPrice { get; set; }
         public string StoreOldShip { get; set; }
         [SuppressMessage("ReSharper", "InconsistentNaming")]
         public int StoreShipID { get; set; }
         public string SellOldShip { get; set; }
         [SuppressMessage("ReSharper", "InconsistentNaming")]
         public int SellShipID { get; set; }
-        public int SellPrice { get; set; }
+        public long SellPrice { get; set; }
     }
 }
